requirements.txt
# Default
beautifulsoup4
rank-bm25
fastapi
langchain
matplotlib
openai
pre-commit
python-dotenv
ray
seaborn
sentence_transformers
slack_bolt
streamlit
structlog
typer
tiktoken

# Vector DB
asyncpg
pgvector
psycopg[binary,pool]
psycopg2-binary
sqlalchemy[asyncio]

# Styling
black
flake8
Flake8-pyproject
isort
pyupgrade

uvicorn
psycopg

